--- src/main/kotlin/com/github/goldsubmarine/restfulhelper/MyBundle.kt
@@ -1,21 +1,0 @@
-package com.github.goldsubmarine.restfulhelper
-
-import com.intellij.AbstractBundle
-import org.jetbrains.annotations.NonNls
-import org.jetbrains.annotations.PropertyKey
-
-@NonNls
-private const val BUNDLE = "messages.MyBundle"
-
-object MyBundle : AbstractBundle(BUNDLE) {
-
-    @Suppress("SpreadOperator")
-    @JvmStatic
-    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
-        getMessage(key, *params)
-
-    @Suppress("SpreadOperator")
-    @JvmStatic
-    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
-        getLazyMessage(key, *params)
-}
