--- src/main/kotlin/com/github/goldsubmarine/restfulhelper/annotations/jaxrs/HEAD.kt
@@ -1,12 +1,0 @@
-package com.github.goldsubmarine.restfulhelper.annotations.jaxrs
-
-import com.intellij.psi.PsiAnnotation
-
-class HEAD(psiAnnotation: PsiAnnotation) : JaxRsMappingAnnotation(psiAnnotation) {
-
-    override fun extractMethod() = METHOD
-
-    companion object {
-        private const val METHOD = "HEAD"
-    }
-}
