--- src/main/kotlin/com/github/goldsubmarine/restfulhelper/annotations/jaxrs/PATCH.kt
@@ -1,12 +1,0 @@
-package com.github.goldsubmarine.restfulhelper.annotations.jaxrs
-
-import com.intellij.psi.PsiAnnotation
-
-class PATCH(psiAnnotation: PsiAnnotation) : JaxRsMappingAnnotation(psiAnnotation) {
-
-    override fun extractMethod() = METHOD
-
-    companion object {
-        private const val METHOD = "PATCH"
-    }
-}
