--- src/main/kotlin/com/github/goldsubmarine/restfulhelper/annotations/micronaut/Put.kt
@@ -1,12 +1,0 @@
-package com.github.goldsubmarine.restfulhelper.annotations.micronaut
-
-import com.intellij.psi.PsiAnnotation
-
-class Put(psiAnnotation: PsiAnnotation) : MicronautMappingAnnotation(psiAnnotation) {
-
-    override fun extractMethod() = METHOD
-
-    companion object {
-        private const val METHOD = "PUT"
-    }
-}
